Attach old databases at upgrade time so upgrade goes through properly

+++ SRBase/src/com/gmail/emerssso/srbase/database/SRTable.java
@@ -49,11 +49,15 @@ public class SRTable {
 	
 	public static void onUpgrade(SQLiteDatabase database, int oldVersion,
 			int newVersion) {
-		Log.w(SRTable.class.getName(), "Upgrading SR database table from version "
-				+ oldVersion + " to " + newVersion);
-		database.execSQL(
-				"INSERT INTO SRdatabase.SR SELECT * FROM SRtable.SR");
-		database.execSQL("DROP TABLE SRtable.SR");
+		if(oldVersion == 1 && newVersion == 2) {
+			Log.w(SRTable.class.getName(), "Upgrading SR database table from version "
+					+ oldVersion + " to " + newVersion);
+			SRTable.onCreate(database);
+			database.execSQL("ATTACH \"/data/data/com.gmail.emerssso.srbase/databases/SRtable.db\" AS SRtable");
+			database.execSQL(
+					"INSERT INTO SRdatabase.SR SELECT * FROM SRtable.SR");
+			database.execSQL("DROP TABLE SRtable.SR");
+		}
 	}
 
 }
